requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.46
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.6
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.17.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-cloud-vision==3.7.2
google-generativeai==0.8.5
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
layoutparser==0.3.4
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
modelscope==1.31.0
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.7.3
paddlepaddle==2.6.1
paddlex==3.3.6
pandas==2.3.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
portalocker==3.2.0
premailer==3.10.0
prettytable==3.16.0
proto-plus==1.26.1
protobuf==3.20.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
pyngrok==7.4.1
pyparsing==3.2.5
pypdfium2==5.0.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.14.3
rarfile==4.2
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
termcolor==3.2.0
tifffile==2025.5.10
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
visualdl==2.5.3
wcwidth==0.2.14
Werkzeug==3.1.3
